requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.1.7
dnspython==2.7.0
eventlet==0.38.0
Flask==3.1.0
Flask-SocketIO==5.4.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pynmeagps==1.0.43
pyrtcm==1.1.2
pyserial==3.5
python-engineio==4.10.1
python-socketio==5.11.4
pyubx2==1.2.48
simple-websocket==1.1.0
Werkzeug==3.1.3
wsproto==1.2.0
pytest==7.4.2
